custom flags moved to a separate file
